abort mpu

+++ gateway/operations/deleteobject.go
@@ -1,6 +1,7 @@
 package operations
 
 import (
+	"net/http"
 	"treeverse-lake/gateway/errors"
 	"treeverse-lake/gateway/permissions"
 )
@@ -15,9 +16,32 @@ func (controller *DeleteObject) GetPermission() string {
 	return permissions.PermissionManageRepos
 }
 
+func (controller *DeleteObject) HandleAbortMultipartUpload(o *PathOperation) {
+	query := o.Request.URL.Query()
+	uploadId := query.Get(QueryParamUploadId)
+
+	// ensure this
+
+	err := o.MultipartManager.Abort(o.Repo, o.Path, uploadId)
+	if err != nil {
+		o.Log().WithError(err).Error("could not abort multipart upload")
+		o.EncodeError(errors.Codes.ToAPIErr(errors.ErrInternalError))
+		return
+	}
+
+	// done.
+	o.ResponseWriter.WriteHeader(http.StatusNoContent)
+}
+
 func (controller *DeleteObject) Handle(o *PathOperation) {
-	// TODO: check if this is an AbortMultipartUpload request
-	// TODO: check if this is a
+	query := o.Request.URL.Query()
+
+	_, hasUploadId := query[QueryParamUploadId]
+	if hasUploadId {
+		controller.HandleAbortMultipartUpload(o)
+		return
+	}
+
 	err := o.Index.DeleteObject(o.Repo, o.Branch, o.Path)
 	if err != nil {
 		o.Log().WithError(err).Error("could not delete key")
